+++ src/relstorage/adapters/postgresql/mover.py
@@ -215,7 +215,9 @@ class PostgreSQLObjectMover(AbstractObjectMover):
             AND tid = %s
         ORDER BY chunk_num
         """
-
+        # Beginning in RelStorage 3, we no longer chunk blobs.
+        # But there may be older blobs already chunked.
+        # TODO: Write a migration to combine chunked blobs.
         f = None
         bytecount = 0
         read_chunk_size = self.blob_chunk_size
@@ -251,10 +253,6 @@ class PostgreSQLObjectMover(AbstractObjectMover):
             f.close()
         return bytecount
 
-    # PostgreSQL < 9.3 only supports up to 2GB of data per BLOB.
-    # Even above that, we can only use larger blobs on 64-bit builds.
-    postgresql_blob_chunk_maxsize = 1 << 31
-
     @metricmethod_sampled
     def upload_blob(self, cursor, oid, tid, filename):
         """Upload a blob from a file.
@@ -289,46 +287,36 @@ class PostgreSQLObjectMover(AbstractObjectMover):
             VALUES (%(oid)s, %(chunk_num)s, %(loid)s)
             """
 
-        blob = None
+        # Since we only run on 9.6 and above, the sizes of large objects
+        # are allowed to exceed 2GB (int_32). The server is already chunking
+        # large objects internally by itself into 4KB pages, so there's no
+        # advantage to us also adding a layer of chunking.
+        #
+        # As long as we keep our usage simple, that's fine. Only
+        # blob.seek(), blob.truncate() and blob.tell() have a need to
+        # use a specific 64-bit function. `export()` and `import()`
+        # (called implicitly by creating the lobject with a local
+        # filename in psycopg2) work with small fixed buffers (8KB) and
+        # don't care about filesize or offset; they just need the
+        # `open` and `read` syscalls to handle 64-bit files (and don't
+        # they have to for Python to handle 64-bit files?)
+        #
+        # psycopg2 explicitly uses the 64 family of functions;
+        # psycopg2cffi does *not* but if it's built on 64-bit
+        # platform, that's fine. pg8000 uses the SQL interfaces, not
+        # the libpq interfaces, and that's also fine. Since we don't use
+        # any of the functions that need 64-bit aware, none of that should be an
+        # issue.
 
-        maxsize = self.postgresql_blob_chunk_maxsize
-        filesize = os.path.getsize(filename)
-        write_chunk_size = self.blob_chunk_size
+        # Create and upload the blob, getting a large object identifier.
+        blob = cursor.connection.lobject(0, 'wb', 0, filename)
+        blob.close()
 
-        if filesize <= maxsize:
-            # File is small enough to fit in one chunk, just use
-            # psycopg2 native file copy support
-            blob = cursor.connection.lobject(0, 'wb', 0, filename)
-            blob.close()
-            params = dict(oid=oid, chunk_num=0, loid=blob.oid)
-            if use_tid:
-                params['tid'] = tid
-            cursor.execute(insert_stmt, params)
-            return
-
-        # We need to divide this up into multiple chunks
-        f = open(filename, 'rb')
-        try:
-            chunk_num = 0
-            while True:
-                blob = cursor.connection.lobject(0, 'wb')
-                params = dict(oid=oid, chunk_num=chunk_num, loid=blob.oid)
-                if use_tid:
-                    params['tid'] = tid
-                cursor.execute(insert_stmt, params)
-
-                for _i in xrange(maxsize // write_chunk_size):
-                    write_chunk = f.read(write_chunk_size)
-                    if not blob.write(write_chunk):
-                        # EOF.
-                        return
-                if not blob.closed:
-                    blob.close()
-                chunk_num += 1
-        finally:
-            f.close()
-            if blob is not None and not blob.closed:
-                blob.close()
+        # Now put it into our blob_chunk table.
+        params = dict(oid=oid, chunk_num=0, loid=blob.oid)
+        if use_tid:
+            params['tid'] = tid
+        cursor.execute(insert_stmt, params)
 
     def store_temps(self, cursor, state_oid_tid_iter):
         # History-preserving storages need the md5 to compare states.