+++ src/relstorage/tests/util.py
@@ -256,6 +256,9 @@ class AbstractTestSuiteBuilder(ABC):
 
         for shared_blob_dir in shared_blob_dir_choices:
             for keep_history in (False, True):
+                # TODO: Make any of the tests that are needing this
+                # subclass StorageCreatingMixin so we unify where
+                # that's handled.
                 def create_storage(name, blob_dir,
                                    shared_blob_dir=shared_blob_dir,
                                    keep_history=keep_history, **kw):